change license from GPLv3 to MIT
Intellectual property does not exist outside of ones mind. Ideas, words and code are not scarce, I cannot be the ministery of truth of the book 1984 to block someoneelese's expression until knowledge becomes corruptible and suscpetible to become a deceiving technique making the masses ignorant. The only reason this project is license is because on a mad man's mind, if I don't license it, it becomes only owner by me, even if publicly available. Reason of the change and choice: MIT is direct, short and easily readable. Does not enforce or restrict anything. GPL was chosen to keep the code open source, but there is no reason someone would use a tor related code that is not open source in the case someone releases it under a more restrictive license. MIT was also chosen to not contaminate a possibly future binary that could make it all GPL.